# swagger.json
{"openapi":"3.0.0","paths":{"/order":{"get":{"operationId":"OrderController_getAllOrders","parameters":[{"name":"clientId","required":true,"in":"query","schema":{"type":"string"}},{"name":"paymentMethod","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"Get all orders","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/"}}}}}},"tags":["Order"]},"post":{"operationId":"OrderController_createOrder","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderVM"}}}}},"tags":["Order"]}},"/order/{id}":{"put":{"operationId":"OrderController_updateOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrderDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderVM"}}}},"404":{"description":"Order not found"}},"tags":["Order"]},"delete":{"operationId":"OrderController_deleteOrder","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Order deleted successfully"},"404":{"description":"Order not found"}},"tags":["Order"]}}},"info":{"title":"Tech Xpress API","description":"The Tech Xpress API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateOrderDto":{"type":"object","properties":{"amount":{"type":"number","description":"Order amount","example":99.99},"longitude":{"type":"number","description":"Longitude coordinate","example":-74.006},"latitude":{"type":"number","description":"Latitude coordinate","example":40.7128},"clientId":{"type":"string","description":"Client ID","example":"123e4567-e89b-12d3-a456-426614174000"},"paymentMethod":{"type":"string","enum":["CASH","VISA"],"description":"Payment method for the order"}},"required":["amount","longitude","latitude","clientId","paymentMethod"]},"OrderVM":{"type":"object","properties":{"id":{"type":"string"},"amount":{"type":"number"},"longitude":{"type":"number"},"latitude":{"type":"number"},"clientId":{"type":"string"},"paymentMethod":{"type":"string"}},"required":["id","amount","longitude","latitude","clientId","paymentMethod"]},"UpdateOrderDto":{"type":"object","properties":{"amount":{"type":"number","description":"Order amount","example":99.99},"longitude":{"type":"number","description":"Longitude coordinate","example":-74.006},"latitude":{"type":"number","description":"Latitude coordinate","example":40.7128},"paymentMethod":{"type":"string","enum":["CASH","VISA"],"description":"Payment method for the order"}},"required":["amount","longitude","latitude","paymentMethod"]}}}}